% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{as_pdf}
\alias{as_pdf}
\title{pdf wrapper that evaluates an arbitrary expression.}
\usage{
as_pdf(fname, w, h, expr, TEST = FALSE, bg = "white")
}
\arguments{
\item{fname}{file path}

\item{w}{the requested width of the PDF}

\item{h}{the requested height of the PDF}

\item{expr}{the expression to evaluate to produce the plot}

\item{TEST}{if FALSE (the default) print to PDF, otherwise print to the current graphics device}

\item{bg}{set the background color of the PDF, defaults to 'white'}
}
\value{
the output of the resulting expression (the plot is likely produced by side effect but may additionally 'return' a value)
}
\description{
pdf wrapper that evaluates an arbitrary expression.
}
