% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{as_title}
\alias{as_title}
\alias{as_title.formula}
\alias{as_title.default}
\alias{as_title.fres}
\alias{as_title.tres}
\title{Make nice plot titles}
\usage{
as_title(x, ...)

\method{as_title}{formula}(x, ...)

\method{as_title}{default}(
  x,
  capitalize_all = TRUE,
  excluded = c("is", "are", "vs", "v.s.", "from", "of", "be", "for", "over"),
  ...
)

\method{as_title}{fres}(x, ...)

\method{as_title}{tres}(x, ...)
}
\arguments{
\item{x}{title content}

\item{...}{pass to other methods}

\item{capitalize_all}{make the first letter cap}

\item{excluded}{exclude making first letter cap, numeric vector}
}
\description{
Make nice plot titles
}
\examples{
\dontrun{
# Display normal title
plot_clean(1,1, main = as_title('this is title'))

# Only capitalize the first character
plot_clean(1,1, main = as_title('this is title', capitalize_all = F))

# What if title is a formula
plot_clean(1,1, main = as_title(p[value](beta[1]) < ~.(0.01) ~ ' is significant'))

# Display t-statistics
plot_clean(1,1)
t_stat = get_t(rnorm(10))
title(main=as_title(t_stat))
}
}
