% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{cbind_list}
\alias{cbind_list}
\title{Make cbind magrittr compatible
(stable)}
\usage{
cbind_list(ll)
}
\arguments{
\item{ll}{the list to bind, each element will be a column}
}
\description{
Make cbind magrittr compatible
(stable)
}
