% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{clip_x}
\alias{clip_x}
\title{Ensure Data Are Within Some Bounds

(stable)}
\usage{
clip_x(x, lim)
}
\arguments{
\item{x}{data}

\item{lim}{clip range, length of 1 or more. If length(lim) is 1, then the clip is symmetric}
}
\description{
Ensure Data Are Within Some Bounds

(stable)
}
