% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{create_frames}
\alias{create_frames}
\title{add decoration (frames) to plots based on their layout location}
\usage{
create_frames(
  layout,
  common = NULL,
  bottom = NULL,
  left = NULL,
  top = NULL,
  right = NULL,
  env = parent.frame(),
  exclude = NULL
)
}
\description{
add decoration (frames) to plots based on their layout location
}
