% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{dir_signature}
\alias{dir_signature}
\title{Create directory signature}
\usage{
dir_signature(
  path = ".",
  file.info = FALSE,
  md5sum = TRUE,
  recursive = TRUE,
  ...
)
}
\arguments{
\item{path}{directory path to generate signatures}

\item{file.info, md5sum, recursive, ...}{passed to \code{\link{fileSnapshot}}}
}
\description{
Create directory signature
}
