% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{do_if}
\alias{do_if}
\title{Clauses With Side Effects (Plotting Etc)

(questioning)}
\usage{
do_if(boolean_expression, if_clause, else_clause = NULL)
}
\arguments{
\item{boolean_expression}{expression that returns true or false}

\item{if_clause}{if true, do if_clause}

\item{else_clause}{if false, do else_clause}
}
\description{
Clauses With Side Effects (Plotting Etc)

(questioning)
}
