% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{do_poly}
\alias{do_poly}
\title{Polygon plot}
\usage{
do_poly(x, y, col, alpha = 50, border = NA, ...)
}
\arguments{
\item{x, y}{x and y}

\item{col}{color}

\item{alpha}{0-255 transparency}

\item{border}{border of polygon}

\item{...}{passed to polygon}
}
\description{
Polygon plot
}
