% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ebar_polygon}
\alias{ebar_polygon}
\title{Draw Symmetric Error Bars

(stable)}
\usage{
ebar_polygon(
  x,
  y,
  sem,
  alpha = 100,
  col = "black",
  fill = col,
  stroke = col,
  border = NA,
  add_line = TRUE,
  lwd = 1,
  ...
)
}
\arguments{
\item{x, y}{plot data}

\item{sem}{error bar half width}

\item{alpha}{transparency}

\item{col}{color}

\item{fill}{error bar color}

\item{stroke}{if add lines, line color}

\item{border}{errorbar border width}

\item{add_line}{add line(x,y) or not}

\item{lwd}{line weight}

\item{...}{passed to \code{\link{lines}}}
}
\description{
Draw Symmetric Error Bars

(stable)
}
\examples{
\dontrun{
plot_clean(0:10, -1:5, xlab = 'X')
ruta_axis(side = 2, at = -1:8)
ebar_polygon(1:10, (1:10)/2, rnorm(10))
}
}
