% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{fix_pdf_name}
\alias{fix_pdf_name}
\title{Ensure that the file names ends in ".pdf"}
\usage{
fix_pdf_name(fname)
}
\arguments{
\item{fname}{potential file name}
}
\description{
Ensure that the file names ends in ".pdf"
}
