% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{format_stat}
\alias{format_stat}
\title{helper function to build value labels}
\usage{
format_stat(nm, stats = c("b", "t", "p"))
}
\arguments{
\item{nm}{statistics}

\item{stats}{names of statistics}
}
\description{
(questioning)
}
