% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_data_range}
\alias{get_data_range}
\title{Get Data Range From A Collection Of Named Lists

(questioning)}
\usage{
get_data_range(ll, name = "range", ...)
}
\arguments{
\item{ll}{list}

\item{name}{element name}

\item{...}{additional params for \link{get_list_elements}}
}
\description{
Get Data Range From A Collection Of Named Lists

(questioning)
}
