% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{get_filename}
\alias{get_filename}
\title{Get the file name from a full file path}
\usage{
get_filename(full_path, keep_extension = FALSE)
}
\arguments{
\item{full_path}{path to file}

\item{keep_extension}{whether the file extension should be retained (defaults to FALSE)}
}
\description{
Get the file name from a full file path
}
