% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{\%&\%}
\alias{\%&\%}
\title{Pipe Function To Paste Two Characters}
\usage{
x \%&\% y
}
\arguments{
\item{x}{character}

\item{y}{character}
}
\value{
paste0(x,y)
}
\description{
Pipe Function To Paste Two Characters
}
