% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{jitr}
\alias{jitr}
\title{Return Jittered X

(experimental)}
\usage{
jitr(x, len = length(x), r)
}
\arguments{
\item{x}{data}

\item{len}{length of x}

\item{r}{jitter size}
}
\description{
Return Jittered X

(experimental)
}
