% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{get_f}
\alias{get_f}
\alias{format_f}
\title{Get statistics from linear model}
\usage{
get_f(formula, ...)

format_f(lm.mod, test_name = "All")
}
\arguments{
\item{formula, ...}{passed to \link{lm}}

\item{lm.mod}{linear model returned by lm()}

\item{test_name}{name}
}
\description{
(questioning)
}
