% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{m_sd}
\alias{m_sd}
\title{Function To Return Mean And Standard Deviation (Na Ignored by default)

(stable)}
\usage{
m_sd(x, na.rm = TRUE)
}
\arguments{
\item{x}{data}

\item{na.rm}{remove NAs?}
}
\description{
Function To Return Mean And Standard Deviation (Na Ignored by default)

(stable)
}
