% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{m_se}
\alias{m_se}
\alias{mat_m_se}
\title{Function To Return Mean And Standard Error

(stable)}
\usage{
m_se(x)

mat_m_se(m, DIM = 2)
}
\arguments{
\item{x}{data}

\item{m}{matrix data}

\item{DIM}{margin, 1 for row, 2 for column. See \link{apply}}
}
\description{
Function To Return Mean And Standard Error

(stable)
}
