% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{NOOP}
\alias{NOOP}
\title{Apply expression but returns something else}
\usage{
NOOP(x, expr = NULL)
}
\arguments{
\item{x}{object to return}

\item{expr}{expression to run}
}
\description{
Evaluate an expression, but then return the input
}
