% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{not_NA}
\alias{not_NA}
\title{Return True If Not Na

(stable)}
\usage{
not_NA(x)

not_NA(x)
}
\arguments{
\item{x}{data}
}
\description{
Return True If Not Na

(stable)

make it easier to say not is.na in a pipe'd context
}
