% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_msg}
\alias{plot_msg}
\title{Show A Blank Plot With Messages

(maturing)}
\usage{
plot_msg(main = "No Conditions Specified")
}
\arguments{
\item{main}{the title/msg to show}
}
\description{
Show A Blank Plot With Messages

(maturing)
}
\examples{
\dontrun{
plot_msg("Let's Say Something")
}
}
