% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{plus_minus}
\alias{plus_minus}
\title{Easy Way To Get +/- From A Long Vector}
\usage{
plus_minus(x, d)
}
\arguments{
\item{x}{data}

\item{d}{plus minus value(s)}
}
\description{
Easy Way To Get +/- From A Long Vector
}
