% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{print_summary}
\alias{print_summary}
\title{Print summary and return original input}
\usage{
print_summary(x, FUN = summary, ...)
}
\arguments{
\item{x}{object to print summary}

\item{FUN}{function to print on}

\item{...}{additional parameters passing to \code{FUN}}
}
\description{
output of a function, but return its input to allow chaining
}
