% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{rbind_list}
\alias{rbind_list}
\title{Make rbind magrittr compatible
(stable)}
\usage{
rbind_list(ll)
}
\arguments{
\item{ll}{the list to bind, each element will be a row}
}
\description{
Make rbind magrittr compatible
(stable)
}
