% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper.R
\name{remove_tail}
\alias{remove_tail}
\title{Remove the last k elements from a vector (list)
Returns x (with a warning) if k < 1}
\usage{
remove_tail(x, k = 1)
}
\arguments{
\item{x}{the vector}

\item{k}{the number of items to remove}
}
\description{
Remove the last k elements from a vector (list)
Returns x (with a warning) if k < 1
}
