% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{round_range}
\alias{round_range}
\title{Get A Integer Interval That Contains X

(maturing)}
\usage{
round_range(x)
}
\arguments{
\item{x}{vector/matrix, numeric}
}
\description{
Get A Integer Interval That Contains X

(maturing)
}
\examples{
\dontrun{
# 0 - 11
round_range(0.5:10.5)
}

}
