% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{row_apply_ii}
\alias{row_apply_ii}
\title{Apply function along the first dimension}
\usage{
row_apply_ii(mat, FUN_, ...)
}
\arguments{
\item{mat}{matrix or array}

\item{FUN_}{function to apply to each row}

\item{...}{additional parameters passing to \code{FUN_}}
}
\description{
row applier with an additional index variable, ii
}
