% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{sapply_ii}
\alias{sapply_ii}
\title{Apply each elements with index as inputs}
\usage{
sapply_ii(X, FUN_, simplify = TRUE, USE.NAMES = TRUE, ...)
}
\arguments{
\item{X}{vector or list}

\item{FUN_}{function to apply on \code{X}}

\item{simplify, USE.NAMES, ...}{passed to \code{sapply}}
}
\description{
sapply with an additional index variable, ii
}
