% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{str_collapse}
\alias{str_collapse}
\title{Convert vector into comma-separated string}
\usage{
str_collapse(x, by = ", ", ...)
}
\arguments{
\item{x}{the vector to collapse}

\item{by}{the separating token (default is ', ')}

\item{...}{further arguments for paste0}
}
\value{
a character vector as produced by paste0
}
\description{
Convert vector into comma-separated string
}
