% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{stretch}
\alias{stretch}
\title{Useful For Plotting When You Want To Go A Bit Beyond The Data

(experimental)}
\usage{
stretch(x, pct)
}
\arguments{
\item{x}{data}

\item{pct}{stretch percentage}
}
\description{
Useful For Plotting When You Want To Go A Bit Beyond The Data

(experimental)
}
