% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{get_t}
\alias{get_t}
\title{helper function for t-tests that returns the values wanted by format_stat}
\usage{
get_t(...)
}
\arguments{
\item{...}{passed to \link{t.test}}
}
\description{
(questioning)
}
