% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{trimmed.mean}
\alias{trimmed.mean}
\title{Mean Of Data After Trimmed

(questioning)}
\usage{
trimmed.mean(x, cutoff = 4)
}
\arguments{
\item{x, cutoff}{passed to \link{trim}}
}
\description{
Mean Of Data After Trimmed

(questioning)
}
