% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{which.equal}
\alias{which.equal}
\title{like which.min, but for equality
useful when an expression for x or y is long}
\usage{
which.equal(x, y)
}
\arguments{
\item{x, y}{vectors to compare}
}
\description{
like which.min, but for equality
useful when an expression for x or y is long
}
